faiss-db/requirements.txt
faiss-cpu
packaging
pika
numpy<2
kafka-python